fix: rename to TestComplexTable
